refactor(ui): move SmallRef into Browse3
