+++ weave-js/src/components/PagePanelComponents/Home/Browse3/filters/FilterTagItem.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import {parseRef} from '../../../../../react';
 import {Timestamp} from '../../../../Timestamp';
 import {UserLink} from '../../../../UserLink';
-import {SmallRef} from '../../Browse2/SmallRef';
+import {SmallRef} from '../smallRef/SmallRef';
 import {
   FilterId,
   getFieldLabel,
